+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/VerticalFlow.jsx
@@ -127,7 +127,7 @@ const getRowHeight = (activeBreakpoint) => {
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Icon',
+    displayName: 'Status',
     width: '60px',
     align: alignTypes.CENTER,
   },