+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-compact-interactive-list` examples to provide a meaningful name to the column with icons.
+
 * Added
   * Added examples and tests for `terra-table` to demonstrate auto layout.