+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+## 1.79.0 - (April 4, 2024)
+
 * Changed
+  * Removed `sortIndicator` from the first column of Table With Sections example.
   * Updated `terra-compact-interactive-list` examples to provide a meaningful name to the column with icons.
 
 * Added